X00T: Move to file-based encryption
Since people have been reluctant to encrypt their userdata due to how QTI
keymaster behaves with full disk encryption, this change should do only minimal
impact. FDE support will be dropped completely in the upcoming Android version.

TWRP support is required, however, and that is, as of this time, planned.

Not using v2 mechanism since that requires updated TWRP support which as of this
time, not yet ready. Also, not ICE because devices launched with Android 11 and
newer are no longer allowed to use it (although still possible for Android 10
and below launched devices to continue to use it).

Signed-off-by: Albert I <kras@raphielgang.org>